Refactoring display module
